Fortran updates, Refractive_Index, Complex to double
Many changes
1) changed Complex struct to use doubles instead of floats
2) all bindings updated in order to support this change
3) Added new function Refractive_Index, which returns a Complex struct
4) Fortran: rewrote all functions to use Fortran strings instead of C
strings, with backwards compatibility for C terminated strings

+++ src/crystal_diffraction.c
@@ -29,8 +29,8 @@ THIS SOFTWARE IS PROVIDED BY David Sagan ''AS IS'' AND ANY EXPRESS OR IMPLIED WA
 
 /*-------------------------------------------------------------------------------------------------- */
 
-float c_abs(Complex x) { 
-  float ans = x.re * x.re - x.im * x.im; 
+double c_abs(Complex x) { 
+  double ans = x.re * x.re - x.im * x.im; 
   ans = sqrt(ans);
   return ans; 
 }
